add update section route
